requirements.txt
# Core Flask dependencies
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# Modern OpenAI with TTS and Vision
openai==1.51.0

# File processing dependencies
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21

# OCR for images (optional)
Pillow==10.0.0
pytesseract==0.3.10

# Development and testing
pytest==7.4.2
pytest-flask==1.2.0

# WebSocket support
Flask-SocketIO==5.3.6
python-socketio==5.8.0


# Migration support
Flask-Migrate==4.0.5

# Additional testing dependencies
pytest-cov>=4.1.0
pytest-mock>=3.11.1
coverage>=7.3.0
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
safety>=2.3.0
bandit>=1.7.0
